requirements.txt
requests
pandas
numpy
mysql-connector-python
python-dotenv